requirements.txt
# Core dependencies
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0
pytorch-lightning>=2.0.0
pytorchvideo>=0.1.5

# Data processing
av>=10.0.0
decord>=0.6.0
hydra-core>=1.3.0
PyYAML>=6.0
h5py>=3.7.0
opencv-python>=4.6.0
pillow>=9.3.0

# Optimization and utilities
rich>=12.6.0
tqdm>=4.64.1
tensorboard>=2.11.0
einops>=0.6.0
omegaconf>=2.3.0
timm>=0.6.12
ninja>=1.11.0

# M1 optimization
--pre torch torchvision torchaudio --extra-index-url https://download.pytorch.org/whl/nightly/cpu